Avoid deadlock condition in auich_read_codec() and auich_write_codec()
When running this code in userspace (e.g. RUMP), it can be permanently
interrupted at any time if the process which contains it is killed.

If this happens in certain parts of auich_read_codec() or auich_write_codec()
routines, ICH_CAS register may be left in an inconsistent state which
leads to deadlock:

  - a command is expected by ICH_CAS before granting access to the driver
  - the driver has requested access and is polling ICH_CAS before issuing
    a command

This commit makes these routines more permissive when run for first time,
so that they can recover from this inconsistent state.

For details on ICH_CAS semaphore, refer to Intel datasheet (page 328):

http://www.intel.com/design/chipsets/datashts/290655.htm